+++ components/calendar/index.ts
@@ -8,4 +8,4 @@ export { calendarProps } from './props'
 export type CalendarExposed = ComponentPublicInstance & InstanceType<typeof Calendar>
 
 export type { CalendarProps, CalendarCProps } from './props'
-export type { WeekIndex, MonthIndex } from './symbol'
+export type { WeekIndex, MonthIndex, CalendarSlots, CalendarPanelSlots } from './symbol'