+++ components/date-picker/date-panel.vue
@@ -2,8 +2,9 @@
 import { Button } from '@/components/button'
 import { CalendarPanel } from '@/components/calendar-panel'
 import { Icon } from '@/components/icon'
+import { ResizeObserver } from '@/components/resize-observer'
 
-import { computed, onMounted, ref, watch } from 'vue'
+import { computed, onMounted, reactive, ref, watch } from 'vue'
 
 import TimeWheel from './time-wheel.vue'
 import { useIcons, useNameHelper } from '@vexip-ui/config'
@@ -16,6 +17,7 @@ import type { MonthIndex } from '@/components/calendar'
 import type { LocaleConfig } from '@vexip-ui/config'
 import type { Dateable } from '@vexip-ui/utils'
 import type {
+  DatePanelSlots,
   DatePickerType,
   DateShortcut,
   DateShortcutsPlacement,
@@ -138,6 +140,8 @@ const emit = defineEmits([
   'time-change'
 ])
 
+defineSlots<DatePanelSlots>()
+
 const nh = useNameHelper('date-picker')
 const icons = useIcons()
 const { isRtl } = useRtl()
@@ -145,7 +149,7 @@ const { isRtl } = useRtl()
 const today = toDate(props.today)
 const monthRange = rangeNumbers(12, 1, 1)
 
-const currentPane = ref<DateType>(
+const currentPanel = ref<DateType>(
   props.type === 'year' ? 'year' : props.type === 'month' ? 'month' : 'date'
 )
 const calendarYear = ref(today.getFullYear())
@@ -154,6 +158,22 @@ const hoveredYear = ref(0) // 0 is no hover (falsy)
 const hoveredMonth = ref(0) // 0 is no hover (falsy)
 const yearRange = ref<number[]>([])
 
+const shortcutsRect = reactive({ width: 0, height: 0 })
+
+const panelStyle = computed(() => {
+  const { width, height } = shortcutsRect
+
+  switch (props.shortcutsPlacement) {
+    case 'top':
+      return { paddingTop: `${height}px` }
+    case 'right':
+      return { paddingRight: `${width}px` }
+    case 'bottom':
+      return { paddingBottom: `${height}px` }
+    default:
+      return { paddingLeft: `${width}px` }
+  }
+})
 const startActivated = computed(() => {
   const activated = props.startActivated
 
@@ -164,7 +184,6 @@ const endActivated = computed(() => {
 
   return activated.year && activated.month && activated.date
 })
-
 const isDatetime = computed(() => {
   return props.type === 'datetime'
 })
@@ -213,8 +232,8 @@ function getMonthLabel(index: number) {
   return props.locale[`month${index as MonthIndex}`]
 }
 
-function togglePane(type: DateType) {
-  currentPane.value = type
+function togglePanel(type: DateType) {
+  currentPanel.value = type
 }
 
 function adjustCalendar(type: 'year' | 'month', amount: number) {
@@ -251,7 +270,7 @@ function handleSelectYear(year: number) {
   calendarYear.value = year
 
   if (props.type !== 'year') {
-    togglePane('month')
+    togglePanel('month')
   } else {
     emitChange([year, 1, 1])
   }
@@ -263,7 +282,7 @@ function handleSelectMonth(month: number) {
   calendarMonth.value = month
 
   if (props.type !== 'month') {
-    togglePane('date')
+    togglePanel('date')
   } else {
     emitChange([calendarYear.value, month, 1])
   }
@@ -282,14 +301,14 @@ function handleEndTimeChange(type: TimeType, time: number) {
 }
 
 function handleDoublePrevClick() {
-  if (currentPane.value === 'year') {
+  if (currentPanel.value === 'year') {
     yearRange.value = rangeNumbers(12, yearRange.value[0] - 10, 1)
   } else {
     adjustCalendar('year', -1)
   }
 }
 function handleDoubleNextClick() {
-  if (currentPane.value === 'year') {
+  if (currentPanel.value === 'year') {
     yearRange.value = rangeNumbers(12, yearRange.value[10], 1)
   } else {
     adjustCalendar('year', 1)
@@ -481,6 +500,18 @@ function refreshCalendar(valueType: 'start' | 'end') {
     calendarMonth.value = props.endActivated.month ? props.endValue.month : today.getMonth() + 1
   }
 }
+
+function handleShortcutsResize(entry: ResizeObserverEntry) {
+  const box = entry.borderBoxSize?.[0]
+
+  if (box) {
+    shortcutsRect.width = box.inlineSize
+    shortcutsRect.height = box.blockSize
+  } else {
+    shortcutsRect.width = entry.contentRect.width
+    shortcutsRect.height = entry.contentRect.height
+  }
+}
 </script>
 
 <template>
@@ -491,27 +522,29 @@ function refreshCalendar(valueType: 'start' | 'end') {
         shortcuts.length && (shortcutsPlacement === 'top' || shortcutsPlacement === 'bottom')
     }"
     :aria-labelledby="labeledBy"
+    :style="panelStyle"
     @click="handleClick"
   >
-    <div
-      v-if="shortcuts.length"
-      :class="[
-        nh.be('list'),
-        nh.bem('list', 'sub'),
-        nh.be('shortcuts'),
-        nh.bem('shortcuts', shortcutsPlacement)
-      ]"
-    >
+    <ResizeObserver v-if="shortcuts.length" :on-resize="handleShortcutsResize">
       <div
-        v-for="(item, index) in shortcuts"
-        :key="index"
-        :class="nh.be('shortcut')"
-        :title="item.name"
-        @click="handleShortcut(index)"
+        :class="[
+          nh.be('list'),
+          nh.bem('list', 'sub'),
+          nh.be('shortcuts'),
+          nh.bem('shortcuts', shortcutsPlacement)
+        ]"
       >
-        {{ item.name }}
+        <div
+          v-for="(item, index) in shortcuts"
+          :key="index"
+          :class="nh.be('shortcut')"
+          :title="item.name"
+          @click="handleShortcut(index)"
+        >
+          {{ item.name }}
+        </div>
       </div>
-    </div>
+    </ResizeObserver>
     <div :class="nh.be('list')" role="application">
       <div :class="nh.be('panel-body')">
         <div :class="nh.be('date-panel')">
@@ -520,31 +553,40 @@ function refreshCalendar(valueType: 'start' | 'end') {
               <Icon v-bind="isRtl ? icons.anglesRight : icons.anglesLeft"></Icon>
             </div>
             <div
-              v-show="currentPane === 'date'"
+              v-show="currentPanel === 'date'"
               :class="[nh.be('arrow'), nh.be('prev-month')]"
               @click="adjustCalendar('month', -1)"
             >
               <Icon v-bind="isRtl ? icons.angleRight : icons.angleLeft"></Icon>
             </div>
             <div :class="nh.be('year-month')">
-              <div key="year" :class="nh.be('year')" @click.stop="togglePane('year')">
-                <template v-if="currentPane === 'year'">
-                  {{ `${yearRange[0]}${locale.year} - ${yearRange[9]}${locale.year}` }}
-                </template>
-                <template v-else>
-                  {{ `${calendarYear}${locale.year}` }}
-                </template>
-              </div>
-              <div
-                v-show="currentPane === 'date'"
-                :class="nh.be('month')"
-                @click.stop="togglePane('month')"
+              <slot
+                name="title"
+                :panel="currentPanel"
+                :year-start="yearRange[0]"
+                :year="calendarYear"
+                :month="calendarMonth"
+                :toggle-panel="togglePanel"
               >
-                {{ getMonthLabel(calendarMonth) }}
-              </div>
+                <div key="year" :class="nh.be('year')" @click.stop="togglePanel('year')">
+                  <template v-if="currentPanel === 'year'">
+                    {{ `${yearRange[0]}${locale.year} - ${yearRange[9]}${locale.year}` }}
+                  </template>
+                  <template v-else>
+                    {{ `${calendarYear}${locale.year}` }}
+                  </template>
+                </div>
+                <div
+                  v-show="currentPanel === 'date'"
+                  :class="nh.be('month')"
+                  @click.stop="togglePanel('month')"
+                >
+                  {{ getMonthLabel(calendarMonth) }}
+                </div>
+              </slot>
             </div>
             <div
-              v-show="currentPane === 'date'"
+              v-show="currentPanel === 'date'"
               :class="[nh.be('arrow'), nh.be('next-month')]"
               @click="adjustCalendar('month', 1)"
             >
@@ -556,7 +598,7 @@ function refreshCalendar(valueType: 'start' | 'end') {
           </div>
           <div :class="nh.be('calendar')">
             <div
-              v-if="currentPane === 'year'"
+              v-if="currentPanel === 'year'"
               :class="nh.be('year-panel')"
               @mouseleave="hoveredYear = 0"
             >
@@ -574,14 +616,23 @@ function refreshCalendar(valueType: 'start' | 'end') {
                 @mouseenter="handleYearHover(item)"
               >
                 <div :class="nh.be('year-label')">
-                  <div :class="nh.be('year-label-inner')">
-                    {{ item }}
-                  </div>
+                  <slot
+                    name="year"
+                    :year="item"
+                    :selected="isSelectedYear(item)"
+                    :is-next="index > 9"
+                    :disabled="isDisabledYear(item)"
+                    :in-range="isYearInRange(item)"
+                  >
+                    <div :class="nh.be('year-label-inner')">
+                      {{ item }}
+                    </div>
+                  </slot>
                 </div>
               </div>
             </div>
             <div
-              v-else-if="currentPane === 'month'"
+              v-else-if="currentPanel === 'month'"
               :class="nh.be('month-panel')"
               @mouseleave="hoveredMonth = 0"
             >
@@ -598,9 +649,18 @@ function refreshCalendar(valueType: 'start' | 'end') {
                 @mouseenter="handleMonthHover(index)"
               >
                 <div :class="nh.be('month-label')">
-                  <div :class="nh.be('month-label-inner')">
-                    {{ getMonthLabel(index) }}
-                  </div>
+                  <slot
+                    name="month"
+                    :year="calendarYear"
+                    :month="index"
+                    :selected="isSelectedMonth(index)"
+                    :disabled="isDisabledMonth(index)"
+                    :in-range="isMonthInRange(index)"
+                  >
+                    <div :class="nh.be('month-label-inner')">
+                      {{ getMonthLabel(index) }}
+                    </div>
+                  </slot>
                 </div>
               </div>
             </div>
@@ -619,7 +679,14 @@ function refreshCalendar(valueType: 'start' | 'end') {
               :week-days="weekDays"
               @select="handleSelectDate"
               @hover="handleHoverDate"
-            ></CalendarPanel>
+            >
+              <template v-if="$slots.week" #week="weekParams">
+                <slot name="week" v-bind="weekParams"></slot>
+              </template>
+              <template v-if="$slots.date" #itemContent="itemParams">
+                <slot name="date" v-bind="itemParams"></slot>
+              </template>
+            </CalendarPanel>
           </div>
         </div>
         <div v-if="isDatetime" :class="nh.be('time-panel')">