+++ components/wheel/symbol.ts
@@ -30,4 +30,8 @@ export interface WheelState {
   decreaseItem(item: ItemState): void
 }
 
+export interface WheelSlots {
+  default?: (params: { option: WheelOption, index: number }) => any
+}
+
 export const WHEEL_STATE: InjectionKey<WheelState> = Symbol('WHEEL_STATE')