+++ docs/demos/config-provider/basis/demo.zh-CN.vue
@@ -9,15 +9,44 @@
     <br />
     <br />
     <Input style="max-width: 400px"></Input>
+    <br />
+    <br />
+    <DatePicker type="month" style="max-width: 400px"></DatePicker>
   </ConfigProvider>
 </template>
 
 <script setup lang="ts">
-import { reactive } from 'vue'
+import { h, reactive } from 'vue'
 
 import type { PropsOptions } from 'vexip-ui'
 
+const months = [
+  '一月',
+  '二月',
+  '三月',
+  '四月',
+  '五月',
+  '六月',
+  '七月',
+  '八月',
+  '九月',
+  '十月',
+  '十一月',
+  '十二月'
+]
 const provideProps = reactive({
+  datePicker: {
+    slots: {
+      panelMonth: ({ month }) =>
+        h(
+          'div',
+          {
+            class: 'vxp-date-picker__month-label-inner'
+          },
+          months[month - 1]
+        )
+    }
+  },
   tag: {
     size: 'large',
     simple: true,