+++ docs/demos/skeleton/loading/demo.zh-CN.vue
@@ -14,6 +14,7 @@
   </Skeleton>
   <Skeleton :loading="loading" activated :repeat="2">
     <div>一些内容</div>
+    <div>一些内容</div>
   </Skeleton>
   <Skeleton :loading="loading" activated width="67%">
     <div>一些内容</div>